Update handler.js
Many users want to be able to ban a specific group or their group from using the bot. I have written the code that works for me and I want to share it with others. There areis one issues: it only works when iI reply to someone in the group, and only the owner can fix it.